Put code quality integration tests into separate packages
