+++ subprojects/code-quality/src/integTest/groovy/org/gradle/api/plugins/quality/checkstyle/CheckstylePluginIntegrationTest.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014 the original author or authors.
+ * Copyright 2018 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,18 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-
-package org.gradle.api.plugins.quality
-
+package org.gradle.api.plugins.quality.checkstyle
 import org.gradle.integtests.fixtures.WellBehavedPluginTest
 
-class FindBugsPluginIntegrationTest extends WellBehavedPluginTest {
-    @Override
-    String getPluginName() {
-        return "findbugs"
-    }
-
+class CheckstylePluginIntegrationTest extends WellBehavedPluginTest {
     @Override
     String getMainTask() {
         return "check"
@@ -32,7 +24,7 @@ class FindBugsPluginIntegrationTest extends WellBehavedPluginTest {
 
     def setup() {
         buildFile << """
-            apply plugin: 'java'
+            apply plugin: 'groovy'
         """
     }
 }
