+++ subprojects/code-quality/src/integTest/groovy/org/gradle/api/plugins/quality/checkstyle/CheckstyleRelocationIntegrationTest.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 the original author or authors.
+ * Copyright 2018 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,33 +14,49 @@
  * limitations under the License.
  */
 
-package org.gradle.api.plugins.quality
+package org.gradle.api.plugins.quality.checkstyle
 
 import org.gradle.integtests.fixtures.AbstractProjectRelocationIntegrationTest
 import org.gradle.test.fixtures.file.TestFile
 
-class PmdRelocationIntegrationTest extends AbstractProjectRelocationIntegrationTest {
+import java.util.regex.Pattern
+
+class CheckstyleRelocationIntegrationTest extends AbstractProjectRelocationIntegrationTest {
 
     @Override
     protected String getTaskName() {
-        return ":pmd"
+        return ":checkstyle"
     }
 
     @Override
     protected void setupProjectIn(TestFile projectDir) {
+        projectDir.file("config/checkstyle/checkstyle.xml") << """<?xml version="1.0"?>
+<!DOCTYPE module PUBLIC
+          "-//Puppy Crawl//DTD Check Configuration 1.3//EN"
+          "http://www.puppycrawl.com/dtds/configuration_1_3.dtd">
+<module name="Checker">
+    <module name="RegexpSingleline">
+       <property name="format" value="\\s+\$"/>
+       <property name="minimum" value="0"/>
+       <property name="maximum" value="0"/>
+       <property name="message" value="Line has trailing spaces."/>
+    </module>
+</module>
+        """
+
         projectDir.file("src/main/java/org/gradle/Class1.java") <<
-            "package org.gradle; class Class1 { public boolean is() { return true; } }"
+            "package org.gradle; class Class1 { public boolean is() { return true; } }  "
         projectDir.file("src/main/java/org/gradle/Class1Test.java") <<
             "package org.gradle; class Class1Test { public boolean is() { return true; } }"
 
         projectDir.file("build.gradle") << """
-            apply plugin: "java"
-            apply plugin: "pmd"
+            apply plugin: "checkstyle"
 
             ${mavenCentralRepository()}
 
-            task pmd(type: Pmd) {
+            task checkstyle(type: Checkstyle) {
                 source "src/main/java"
+                classpath = files()
                 ignoreFailures = true
             }
         """
@@ -48,7 +64,7 @@ class PmdRelocationIntegrationTest extends AbstractProjectRelocationIntegrationT
 
     @Override
     protected extractResultsFrom(TestFile projectDir) {
-        projectDir.file("build/reports/pmd/pmd.xml").text
-            .replaceAll(/timestamp=".*?"/, 'timestamp="[NUMBER]"')
+        return projectDir.file("build/reports/checkstyle/checkstyle.xml").text
+            .replaceAll(Pattern.quote(projectDir.absolutePath + File.separator) + ".*?" + Pattern.quote(File.separator), "")
     }
 }
