+++ subprojects/code-quality/src/integTest/groovy/org/gradle/api/plugins/quality/findbugs/FindBugsClasspathValidationIntegrationTest.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014 the original author or authors.
+ * Copyright 2018 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,7 +15,7 @@
  */
 
 
-package org.gradle.api.plugins.quality
+package org.gradle.api.plugins.quality.findbugs
 
 import org.gradle.integtests.fixtures.AbstractIntegrationSpec
 import org.gradle.util.Requires
