+++ subprojects/code-quality/src/integTest/groovy/org/gradle/api/plugins/quality/findbugs/FindBugsIntegrationTest.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011 the original author or authors.
+ * Copyright 2018 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,18 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradle.api.plugins.quality
-import org.gradle.integtests.fixtures.WellBehavedPluginTest
+package org.gradle.api.plugins.quality.findbugs
 
-class PmdPluginIntegrationTest extends WellBehavedPluginTest {
-    def setup() {
-        buildFile << """
-            apply plugin: "java"
-        """
-    }
-
-    @Override
-    String getMainTask() {
-        return "check"
-    }
+class FindBugsIntegrationTest extends AbstractFindBugsPluginIntegrationTest {
 }
