+++ subprojects/code-quality/src/integTest/groovy/org/gradle/api/plugins/quality/findbugs/FindBugsPluginIntegrationTest.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011 the original author or authors.
+ * Copyright 2018 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,10 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradle.api.plugins.quality
+
+
+package org.gradle.api.plugins.quality.findbugs
+
 import org.gradle.integtests.fixtures.WellBehavedPluginTest
 
-class CheckstylePluginIntegrationTest extends WellBehavedPluginTest {
+class FindBugsPluginIntegrationTest extends WellBehavedPluginTest {
+    @Override
+    String getPluginName() {
+        return "findbugs"
+    }
+
     @Override
     String getMainTask() {
         return "check"
@@ -24,7 +32,7 @@ class CheckstylePluginIntegrationTest extends WellBehavedPluginTest {
 
     def setup() {
         buildFile << """
-            apply plugin: 'groovy'
+            apply plugin: 'java'
         """
     }
 }
