+++ subprojects/code-quality/src/integTest/groovy/org/gradle/api/plugins/quality/findbugs/FindBugsRelocationIntegrationTest.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 the original author or authors.
+ * Copyright 2018 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.api.plugins.quality
+package org.gradle.api.plugins.quality.findbugs
 
 import groovy.xml.XmlUtil
 import org.gradle.integtests.fixtures.AbstractProjectRelocationIntegrationTest
